Inbox Listings, kind_filter, Packages and more (#69)
* Write requests

* Create packages, optimize imports

* Add api_type to request params

* Fix submission.morechildren

Fix `morechildren()` using "json" `api_type` response.

* Make morechildren async generator

Update tests for async generator.

* Create Message model

* Add support for message in listing

* Enabled usage of kind_filter in listing

Ensured usage of iterator method `next()` or `for` syntax to apply `kind_filter` when using `class Listing`.

* Fix imports in message.py

* Implement AuthenticatedUser inbox methods

inbox(), sent() and unread()
