+++ apraw/const.py
@@ -1,5 +1,5 @@
 """aPRAW constants."""
-from .endpoints import API_PATH
+from .endpoints import API_PATH  # noqa: F401
 
 __version__ = "0.4.0"
 __tag__ = "alpha"
