+++ apraw/models/reddit/comment.py
@@ -1,19 +1,19 @@
 from typing import TYPE_CHECKING, AsyncIterator, Dict, List
 
-from .helpers.apraw_base import aPRAWBase
-from .helpers.item_moderation import PostModeration
-from .mixins.author import AuthorMixin
-from .mixins.deletable import DeletableMixin
-from .mixins.hideable import HideableMixin
-from .mixins.savable import SavableMixin
-from .mixins.subreddit import SubredditMixin
-from .mixins.votable import VotableMixin
 from .redditor import Redditor
-from .subreddit import Subreddit
-from ..endpoints import API_PATH
+from ..helpers.apraw_base import aPRAWBase
+from ..helpers.item_moderation import PostModeration
+from ..mixins.author import AuthorMixin
+from ..mixins.deletable import DeletableMixin
+from ..mixins.hideable import HideableMixin
+from ..mixins.savable import SavableMixin
+from ..mixins.subreddit import SubredditMixin
+from ..mixins.votable import VotableMixin
+from ..subreddit.subreddit import Subreddit
+from ...const import API_PATH
 
 if TYPE_CHECKING:
-    from ..reddit import Reddit
+    from ...reddit import Reddit
     from .submission import Submission
