+++ apraw/models/user.py
@@ -3,9 +3,11 @@ from typing import TYPE_CHECKING, Dict, List
 
 import aiohttp
 
-from ..endpoints import API_PATH
 from .helpers.apraw_base import aPRAWBase
-from .redditor import Redditor
+from .helpers.generator import ListingGenerator
+from .helpers.streamable import Streamable
+from .reddit.redditor import Redditor
+from ..endpoints import API_PATH
 
 if TYPE_CHECKING:
     from ..reddit import Reddit
@@ -183,6 +185,18 @@ class AuthenticatedUser(Redditor):
             resp = await self.reddit.get_request(API_PATH["me_karma"])
             self._karma = [Karma(self.reddit, d) for d in resp["data"]]
         return self._karma
+
+    @Streamable.streamable
+    async def inbox(self, *args, **kwargs) -> ListingGenerator:
+        return ListingGenerator(self.reddit, API_PATH["message_inbox"], *args, **kwargs)
+
+    @Streamable.streamable
+    async def sent(self, *args, **kwargs) -> ListingGenerator:
+        return ListingGenerator(self.reddit, API_PATH["message_sent"], *args, **kwargs)
+
+    @Streamable.streamable
+    async def unread(self, *args, **kwargs) -> ListingGenerator:
+        return ListingGenerator(self.reddit, API_PATH["message_unread"], *args, **kwargs)
 
 
 class Karma(aPRAWBase):
