+++ requests/requests.http
@@ -61,3 +61,12 @@ GET {{baseUrl}}/api/v1/me/karma
 User-Agent: {{user_agent}}
 Content-Type: application/json
 Authorization: {{getToken.response.body.token_type}} {{getToken.response.body.access_token}}
+
+
+###
+
+# @name getInbox
+GET {{baseUrl}}/message/inbox
+User-Agent: {{user_agent}}
+Content-Type: application/json
+Authorization: {{getToken.response.body.token_type}} {{getToken.response.body.access_token}}
